Update simulator to allow path lengths over 2^31.
git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@7785 bbc10eb1-c90d-0410-af57-cb519fbb1720

+++ prism/src/cex/CexPathAsBDDs.java
@@ -74,7 +74,7 @@ public class CexPathAsBDDs implements PathFullInfo
 	// ACCESSORS (for PathFullInfo)
 	
 	@Override
-	public int size()
+	public long size()
 	{
 		return states.size() - 1;
 	}
@@ -140,13 +140,13 @@ public class CexPathAsBDDs implements PathFullInfo
 	}
 	
 	@Override
-	public int loopStart()
+	public long loopStart()
 	{
 		return 0;
 	}
 	
 	@Override
-	public int loopEnd()
+	public long loopEnd()
 	{
 		return 0;
 	}
