+++ prism/src/simulator/sampler/SamplerBoundedUntilDisc.java
@@ -65,7 +65,7 @@ public class SamplerBoundedUntilDisc extends SamplerBoolean
 		if (valueKnown)
 			return true;
 		
-		int pathSize = path.size();
+		long pathSize = path.size();
 		// Upper bound exceeded
 		if (pathSize > ub) {
 			valueKnown = true;
