+++ prism/src/userinterface/SimulationInformation.java
@@ -51,7 +51,7 @@ public class SimulationInformation
 	private double width;
 	private double confidence;
 	private int numSamples;
-	private int maxPathLength;
+	private long maxPathLength;
 
 	private boolean distributed;
 	private boolean maxRewardGiven;
@@ -72,7 +72,7 @@ public class SimulationInformation
 		this.width = settings.getDouble(PrismSettings.SIMULATOR_DEFAULT_WIDTH);
 		this.confidence = settings.getDouble(PrismSettings.SIMULATOR_DEFAULT_CONFIDENCE);
 		this.numSamples = settings.getInteger(PrismSettings.SIMULATOR_DEFAULT_NUM_SAMPLES);
-		this.maxPathLength = settings.getInteger(PrismSettings.SIMULATOR_DEFAULT_MAX_PATH);
+		this.maxPathLength = settings.getLong(PrismSettings.SIMULATOR_DEFAULT_MAX_PATH);
 
 
 		this.distributed = false;
@@ -194,7 +194,7 @@ public class SimulationInformation
 	 * Getter for property maxPathLength.
 	 * @return Value of property maxPathLength.
 	 */
-	public int getMaxPathLength()
+	public long getMaxPathLength()
 	{
 		return maxPathLength;
 	}
@@ -203,7 +203,7 @@ public class SimulationInformation
 	 * Setter for property maxPathLength.
 	 * @param maxPathLength New value of property maxPathLength.
 	 */
-	public void setMaxPathLength(int maxPathLength)
+	public void setMaxPathLength(long maxPathLength)
 	{
 		this.maxPathLength = maxPathLength;
 	}
